requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
supabase==2.3.4
psycopg2-binary==2.9.9 